+++ tests/agent/test_agent_react.py
@@ -4,7 +4,12 @@ from inspect_ai import Task, eval
 from inspect_ai.agent._agent import Agent, AgentState, agent
 from inspect_ai.agent._handoff import handoff
 from inspect_ai.agent._react import react
-from inspect_ai.agent._types import AgentAttempts, AgentPrompt, AgentSubmit
+from inspect_ai.agent._types import (
+    DEFAULT_CONTINUE_PROMPT,
+    AgentAttempts,
+    AgentPrompt,
+    AgentSubmit,
+)
 from inspect_ai.dataset import Sample
 from inspect_ai.log import EvalLog
 from inspect_ai.model import ChatMessageUser, ModelOutput, get_model
@@ -319,7 +324,8 @@ def test_react_agent_retries_with_custom_incorrect_message():
 
 
 def test_react_agent_on_continue_str():
-    on_continue = "Please keep going and call the {submit}() tool!"
+    # should always return on the on_continue str
+    on_continue = "Please keep going!"
     addition_task = Task(
         dataset=[Sample(input="What is 1 + 1?", target="2")],
         solver=react(tools=[addition()], on_continue=on_continue),
@@ -341,7 +347,35 @@ def test_react_agent_on_continue_str():
     assert log.status == "success"
     assert log.samples
     messages = log.samples[0].messages
-    assert messages[-2].text == on_continue.format(submit="submit")
+    assert len(messages) == 7
+    assert messages[-2].text == on_continue
+
+
+def test_react_agent_on_continue_null():
+    # should return default str if there are no tool calls
+    addition_task = Task(
+        dataset=[Sample(input="What is 1 + 1?", target="2")],
+        solver=react(tools=[addition()]),
+        scorer=compare_quantities(),
+    )
+    log = eval(
+        addition_task,
+        model=get_model(
+            "mockllm/model",
+            custom_outputs=[
+                ModelOutput.for_tool_call(
+                    "mockllm/model", "addition", {"x": 1, "y": 1}
+                ),
+                ModelOutput.from_content("mockllm/model", "I give up!"),
+                ModelOutput.for_tool_call("mockllm/model", "submit", {"answer": "2"}),
+            ],
+        ),
+    )[0]
+    assert log.status == "success"
+    assert log.samples
+    messages = log.samples[0].messages
+    assert len(messages) == 7  # on continue should only appear once
+    assert messages[-2].text == DEFAULT_CONTINUE_PROMPT.format(submit="submit")
 
 
 def test_react_agent_on_continue_func():